+++ plugin/reader/tdenginereader/src/main/java/com/wgzhao/datax/plugin/reader/tdenginereader/TDengineReader.java
@@ -1,0 +1,104 @@
+package com.wgzhao.datax.plugin.reader.tdenginereader;
+
+import com.wgzhao.datax.common.plugin.RecordSender;
+import com.wgzhao.datax.common.spi.Reader;
+import com.wgzhao.datax.common.util.Configuration;
+import com.wgzhao.datax.plugin.rdbms.reader.CommonRdbmsReader;
+import com.wgzhao.datax.plugin.rdbms.reader.Constant;
+import com.wgzhao.datax.plugin.rdbms.util.DBUtil;
+import com.wgzhao.datax.plugin.rdbms.util.DataBaseType;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+import java.util.Properties;
+
+public class TDengineReader
+        extends Reader
+{
+
+    private static final DataBaseType DATABASE_TYPE = DataBaseType.TDengine;
+
+    public static class Job
+            extends Reader.Job
+    {
+        private static final Logger LOG = LoggerFactory.getLogger(Job.class);
+
+        private Configuration originalConfig = null;
+        private CommonRdbmsReader.Job commonRdbmsReaderJob;
+
+        @Override
+        public void init()
+        {
+            this.originalConfig = getPluginJobConf();
+
+            this.commonRdbmsReaderJob = new CommonRdbmsReader.Job(DATABASE_TYPE);
+            this.commonRdbmsReaderJob.init(this.originalConfig);
+        }
+
+        @Override
+        public void preCheck()
+        {
+            init();
+            this.commonRdbmsReaderJob.preCheck(this.originalConfig, DATABASE_TYPE);
+        }
+
+        @Override
+        public List<Configuration> split(int adviceNumber)
+        {
+            return this.commonRdbmsReaderJob.split(this.originalConfig, adviceNumber);
+        }
+
+        @Override
+        public void post()
+        {
+            this.commonRdbmsReaderJob.post(this.originalConfig);
+        }
+
+        @Override
+        public void destroy()
+        {
+            this.commonRdbmsReaderJob.destroy(this.originalConfig);
+        }
+    }
+
+    public static class Task
+            extends Reader.Task
+    {
+
+        private Configuration readerSliceConfig;
+        private CommonRdbmsReader.Task commonRdbmsReaderTask;
+
+        @Override
+        public void init()
+        {
+            this.readerSliceConfig = getPluginJobConf();
+            this.commonRdbmsReaderTask = new CommonRdbmsReader.Task(DATABASE_TYPE, getTaskGroupId(), getTaskId());
+            this.commonRdbmsReaderTask.init(this.readerSliceConfig);
+        }
+
+        @Override
+        public void startRead(RecordSender recordSender)
+        {
+            // TDengine does not support fetch size
+            int fetchSize = this.readerSliceConfig.getInt(Constant.FETCH_SIZE);
+            this.commonRdbmsReaderTask.startRead(this.readerSliceConfig, recordSender, getTaskPluginCollector(), fetchSize);
+        }
+
+        @Override
+        public void post()
+        {
+            this.commonRdbmsReaderTask.post(this.readerSliceConfig);
+        }
+
+        @Override
+        public void destroy()
+        {
+            this.commonRdbmsReaderTask.destroy(this.readerSliceConfig);
+        }
+    }
+}